+++ main.tf
@@ -9,16 +9,24 @@ terraform {
 
 module "packer-vmis" {
   source = "./packer-vmis"
-  virtual_machine_image_master = var.virtual_machine_image_master
-  private_key_path_master = var.private_key_path_master
-  virtual_machine_image_worker = var.virtual_machine_image_worker
-  private_key_path_worker = var.private_key_path_worker
+  virtual_machine_image = var.virtual_machine_image
+  private_key_path = var.private_key_path 
 }
 
-module "vbox-cluster" {  
-  source = "./vbox-cluster"
-  virtual_machine_image_master = var.virtual_machine_image_master
-  private_key_path_master = var.private_key_path_master
-  virtual_machine_image_worker = var.virtual_machine_image_worker
-  private_key_path_worker = var.private_key_path_worker
+module "kubernetes-cluster" {  
+  source = "./kubernetes-cluster"
+  virtual_machine_image = var.virtual_machine_image
+  private_key_path = var.private_key_path
+  ip_base = var.ip_base
+  cluster_name = var.cluster_name
+  master_count = var.master_count
+  master_cpus = var.master_cpus
+  master_memory = var.master_memory
+  worker_count = var.worker_count
+  worker_cpus = var.worker_cpus
+  worker_memory = var.worker_memory
+
+  #depends_on = [module.packer-vmis]
 }
+
+