+++ outputs.tf
@@ -1,0 +1,4 @@
+output "virtual-machine-instance-outputs" {
+    description = "The outputs from the virtual-machine-instance module."
+    value       = module.kubernetes-cluster
+}